@@ -7,10 +7,10 @@
     "dev": "ts-node .electron-nuxt/dev.ts",
     "build": "node .electron-nuxt/build.js",
     "test": "",
-    "lint": "node_modules/.bin/eslint --ext .js,.vue -f ./node_modules/eslint-friendly-formatter ./src",
+    "lint": "eslint --ext .js,.vue",
     "lint:fix": "yarn run lint -- --fix",
     "preinstall": "node .electron-nuxt/check-engines.js",
-    "postinstall": "electron-builder install-app-deps & yarn run lint:fix"
+    "postinstall": "electron-builder install-app-deps"
   },
   "keywords": [],
   "author": "ry",
@@ -23,36 +23,36 @@
     "@types/node": "12.12.6"
   },
   "devDependencies": {
-    "@mdi/font": "5.6.55",
-    "@nuxtjs/eslint-config": "1.1.2",
-    "@nuxtjs/vuetify": "1.11.2",
+    "@mdi/font": "5.9.55",
+    "@nuxtjs/eslint-config": "5.0.0",
+    "@nuxtjs/vuetify": "1.11.3",
     "@xpda-dev/core": "^0.1.5",
     "@xpda-dev/electron-builder": "^0.2.1",
     "@xpda-dev/electron-launcher": "^0.2.1",
     "@xpda-dev/webpack-step": "^0.1.2",
     "babel-eslint": "10.1.0",
     "browser-env": "3.3.0",
-    "cross-env": "7.0.2",
+    "cross-env": "7.0.3",
     "deepmerge": "4.2.2",
     "electron": "8.3.0",
-    "electron-builder": "22.8.1",
-    "electron-debug": "3.1.0",
-    "eslint": "5.16.0",
-    "eslint-config-prettier": "^7.0.0",
+    "electron-builder": "22.9.1",
+    "electron-debug": "3.2.0",
+    "eslint": "^7.18.0",
+    "eslint-config-prettier": "^7.2.0",
     "eslint-friendly-formatter": "4.0.1",
     "eslint-loader": "4.0.2",
-    "eslint-plugin-prettier": "^3.2.0",
+    "eslint-plugin-prettier": "^3.3.1",
     "native-ext-loader": "2.3.0",
-    "node-sass": "4.14.1",
-    "nuxt": "2.14.5",
+    "node-sass": "5.0.0",
+    "nuxt": "2.14.12",
     "prettier": "^2.2.1",
     "roboto-fontface": "0.10.0",
-    "sass-loader": "10.0.2",
-    "ts-loader": "^8.0.12",
+    "sass-loader": "10.1.1",
+    "ts-loader": "^8.0.14",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.2",
+    "typescript": "^4.1.3",
     "vue-devtools": "5.1.4",
-    "vuetify": "2.3.20",
+    "vuetify": "2.4.3",
     "webpack": "^4.46.0",
     "webpack-node-externals": "2.5.2"
   },
@@ -64,5 +64,6 @@
     "heapdump": "^0.3.15",
     "vue-form": "^4.10.3",
     "vuelidate": "^0.7.6"
-  }
+  },
+  "repository": "git@github.com:anthonyr4093/melee-ach.git"
 }
